medisure-ai-backend-4874b60d9380591b433e899995d5cce6c3773183/deeplearning_text_summarization/requirements.txt
bert-extractive-summarizer
spacy
transformers==2.2.2
neuralcoref
torch